requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
aioschedule>=0.5.2
pytz>=2024.1